Make sure "Shank & Wrot, Orc Scavengers" cannot lead troops
According to the wording of the book it should be impossible to add warriors to the warband of Shank and Wrot;

"Shank & Wrot, Orc Scavengers, is made up of Shank, Wrot and a Snow Troll, and will count as three models. Shank is always the Captain of the Warband with Wrot and the Snow Troll as his Followers; Wrot always counts as an Independent Hero in Shank’s Warband. No other models may be part of Shank’s Warband."